# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Umbrella Industrial Services ERP API",
        "description": "Documentation API de l'ERP Umbrella",
        "contact": {
            "email": "support@umbrella.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/candidatures": {
            "get": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Récupérer toutes les candidatures",
                "operationId": "02ce0996996e5008ffdec64055a93565",
                "responses": {
                    "200": {
                        "description": "Liste des candidatures récupérée avec succès"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Créer une nouvelle candidature",
                "operationId": "c53567ed67a01536c53587ff3cbb0da6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "first_name",
                                    "last_name",
                                    "location",
                                    "email"
                                ],
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Yaoundé"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+237699123456"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Candidature créée avec succès"
                    },
                    "422": {
                        "description": "Validation des champs échouée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/candidatures/{candidature}": {
            "get": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Récupérer une candidature par ID",
                "operationId": "0b3282cd882f4e4f6466f5588a054a9e",
                "parameters": [
                    {
                        "name": "candidature",
                        "in": "path",
                        "description": "ID de la candidature",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Candidature récupérée avec succès"
                    },
                    "404": {
                        "description": "Candidature non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Mettre à jour une candidature",
                "operationId": "05586317a8543d5fbd9aaf107d39775e",
                "parameters": [
                    {
                        "name": "candidature",
                        "in": "path",
                        "description": "ID de la candidature",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Douala"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+237699123456"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "john.doe@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Candidature mise à jour avec succès"
                    },
                    "404": {
                        "description": "Candidature non trouvée"
                    },
                    "422": {
                        "description": "Validation échouée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Supprimer une candidature",
                "operationId": "9ba59d2d1ca0f12887b3deb7f98f507d",
                "parameters": [
                    {
                        "name": "candidature",
                        "in": "path",
                        "description": "ID de la candidature",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Candidature supprimée avec succès"
                    },
                    "404": {
                        "description": "Candidature non trouvée"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/candidatures/{candidature}/validate": {
            "post": {
                "tags": [
                    "Candidatures"
                ],
                "summary": "Valider une candidature et créer un agent",
                "operationId": "cd964def58d3bd2dcac055166ba7a3c5",
                "parameters": [
                    {
                        "name": "candidature",
                        "in": "path",
                        "description": "ID de la candidature",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "site_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Candidature validée et agent créé"
                    }
                }
            }
        },
...
    },
...
}